basic support for docker.connection override on a per image resource. moved image into helper.
