+++ libraries/helpers_container.rb
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 module DockerHelpers
   module Container
     ################
